requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.16.0
webdriver-manager>=4.0.1
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
pandas>=2.1.0